% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelutilsr.R
\name{add_cellstyles_to_wb}
\alias{add_cellstyles_to_wb}
\title{Takes workbook object (\code{wb}), the format list (\code{fmt_list}),
\code{fmt_row}, a nd \code{fmt_col}).}
\usage{
add_cellstyles_to_wb(wb, m_df, sheet, fmt_list, fmt_row, fmt_col)
}
\arguments{
\item{wb}{workbook object}

\item{m_df}{dataframe to receive formatted worksheet}

\item{sheet}{character vector with name of worksheet}

\item{fmt_list}{list of format list(s) used to format the worksheet}

\item{fmt_row}{a matrix of row numbers where the value of every cell
has the row number of that cell.}

\item{fmt_col}{a matrix of column numbers the value of every cell has
the column number of that cell.}
}
\value{
\code{wb}
}
\description{
Takes workbook object (\code{wb}), the format list (\code{fmt_list}),
\code{fmt_row}, a nd \code{fmt_col}).
}
