% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelutilsr.R
\name{create_sheet_if_needed}
\alias{create_sheet_if_needed}
\title{Creates an Excel worksheet if it is needed.}
\usage{
create_sheet_if_needed(wb, m_df, sheet, header, create)
}
\arguments{
\item{wb}{workbook object}

\item{m_df}{dataframe to receive formated worksheet}

\item{sheet}{character vector with name of worksheet}

\item{header}{logical vector of length one having TRUE if a header is to be
created in the worksheet.}

\item{create}{logical vector of length 1 that is TRUE if a worksheet is to
be created if not present.}
}
\description{
Takes a workbook object, a dataframe, the name of a worksheet, whether or
not the sheet is to have a header, and whether or not the sheet is to be
created if it does not exists, and either creates a worksheet or does based
on the values of the arguments and whether or not the sheet already exists.
}

