% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelutilsr.R
\name{set_cellstyle}
\alias{set_cellstyle}
\title{Set the cell styles to be used in a workbook}
\usage{
set_cellstyle(fmt, wb)
}
\arguments{
\item{fmt}{format list(s) used to format the worksheet}

\item{wb}{workbook object}
}
\value{
cell style formats based on fmt
}
\description{
Cell styles are according to the format matrix provided (\code{fmt}.
Creates an anonymous cell style within the fmt_cellstyle object.
}
