% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelutilsr.R
\name{add_formated_worksheet}
\alias{add_formated_worksheet}
\title{Adds formated data from a dataframe to a worksheet in a provided Excel file}
\usage{
add_formated_worksheet(m_df, excel_file, sheet = sheet, header = TRUE,
  fmt_list, create = TRUE)
}
\arguments{
\item{m_df}{dataframe to receive formated worksheet}

\item{excel_file}{character vector of length 1 with file name of Excel
workbook}

\item{sheet}{character vector with name of worksheet}

\item{header}{logical vector of length one having TRUE if a header is to be
created in the worksheet.}

\item{fmt_list}{list of format list used to format worksheet}

\item{create}{logical vector of length one having TRUE if a new worksheet
is to be created}
}
\value{
Excel file name with formatted worksheet.
}
\description{
Takes a dataframe, an Excel file name, and a list of styles to be used one
the worksheet if the user provided function evalutes to TRUE for a cell
within the dataframe.
}
\examples{
library(stringi)
library(XLConnect)

## borders can be any of the following line types
names(XLC[stri_detect_regex(names(XLC), pattern = "^BORDER.")])
## colors can be any of the following
names(XLC[stri_detect_regex(names(XLC), pattern = "^COLOR.")])
## fill_pattern can be any of the following
names(XLC[stri_detect_regex(names(XLC), pattern = "^FILL.")])
my_f1 <- function(x) {x > 7}
fmt_lst_1 <- list(test = my_f1,
                wrap = TRUE,
                fill_pattern = as.integer(XLC$FILL.SOLID_FOREGROUND),
                foreground_color = as.integer(XLC$COLOR.LAVENDER),
                border = list(side = "all",
                              type = as.integer(XLC$BORDER.THICK),
                              color = XLC$COLOR.DARK_BLUE)
                )
my_df <- data.frame(ColA = stri_c("Name_", 1:4),
                    ColB = c(4, 7, 8, 9),
                    ColC = c(100, 300, 3000, 132),
                    ColD = 1:4)
result <- add_formated_worksheet(my_df, "example_wkbk.xlsx",
                                 sheet = "my_test",
                                 header = TRUE,
                                 fmt_list = list(fmt_lst_1), create = TRUE)

my_f2 <- function(x) {(x \%\% 3) == 0}
fmt_lst_2 <- list(test = my_f2,
                wrap = TRUE,
                fill_pattern = as.integer(XLC$FILL.SOLID_FOREGROUND),
                foreground_color = as.integer(XLC$COLOR.TAN),
                border = list(side = "all",
                              type = as.integer(XLC$BORDER.DOUBLE),
                              color = XLC$COLOR.GREY_80_PERCENT)
                )
my_other_df <- data.frame(LETTERS = LETTERS[1:10],
                          NUMBERS = 1:10,
                          FRACTIONAL = 1 / (1:10))
result <- add_formated_worksheet(my_df, "example_wkbk.xlsx",
                                 sheet = "my_test",
                                 header = TRUE,
                                 fmt_list = list(fmt_lst_1, fmt_lst_2),
                                 create = TRUE)
}

