% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelutilsr.R
\name{set_cellstyle}
\alias{set_cellstyle}
\title{Set the cell styles in the workbook according to the format matrix provided.}
\usage{
set_cellstyle(fmt, wb)
}
\value{
cell style formats based on fmt
}
\description{
Creates an anonymous cell style within the fmt_cellstyle object.
}

